add readme file

+++ backend/controllers/task.controller.js
@@ -20,7 +20,8 @@ const updateTask = async (req, res) => {
   const { title } = req.body;
   const { id } = req.params;
 
-  await taskExist(id, res);
+  const taskExists = await taskExist(id, res);
+  if (!taskExists) return;
 
   if (!title || title.trim() == "") {
     return res.status(400).json({ message: "Title can't be empty." });
@@ -34,7 +35,8 @@ const updateTask = async (req, res) => {
 const deleteTask = async (req, res) => {
   const { id } = req.params;
 
-  await taskExist(id, res);
+  const taskExists = await taskExist(id, res);
+  if (!taskExists) return;
 
   await Task.findByIdAndDelete(id);
   res.json({ message: "Task deleted" });
@@ -47,8 +49,10 @@ const taskExist = async (taskId, response) => {
 
   const task = await Task.findById(taskId);
   if (!task) {
-    return response.status(404).json({ message: "Task not found." });
+    response.status(404).json({ message: "Task not found." });
+    return false;
   }
+  return true;
 };
 
 module.exports = { getTasks, createTask, updateTask, deleteTask };
